requirements.txt
httpx==0.14.1
python-dotenv==0.14.0